refactor: extract workspace and turn services
